Consolidate

+++ ci/install_snap_dependencies.sh
@@ -1,10 +1,12 @@
 #!/bin/bash -xe
 
-sudo apt update
+if test "$(uname -s)" = "Linux"; then
+    sudo apt update
 
-mkdir -p fakesnap/snap
-cp ci/snapcraft.yaml fakesnap/snap/
-pushd fakesnap
-snapcraft pull desktop-gtk3 electron-deps
-popd
-rm -rf fakesnap
+    mkdir -p fakesnap/snap
+    cp ci/snapcraft.yaml fakesnap/snap/
+    pushd fakesnap
+    snapcraft pull desktop-gtk3 electron-deps
+    popd
+    rm -rf fakesnap
+fi
